ユーザーのログの出力内容変更

+++ backend/application-core/src/main/java/com/memoblend/applicationcore/applicationservice/UserApplicationService.java
@@ -1,10 +1,15 @@
 package com.memoblend.applicationcore.applicationservice;
 
 import java.util.List;
+import java.util.Locale;
 import java.util.logging.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
+import com.memoblend.applicationcore.constant.MessageIdConstants;
 import com.memoblend.applicationcore.user.User;
+import com.memoblend.applicationcore.user.UserAlreadyExistException;
+import com.memoblend.applicationcore.user.UserDomainService;
 import com.memoblend.applicationcore.user.UserNotFoundException;
 import com.memoblend.applicationcore.user.UserRepository;
 import com.memoblend.systemcommon.constant.SystemPropertyConstants;
@@ -19,16 +24,18 @@ public class UserApplicationService {
 
   @Autowired
   private UserRepository userRepository;
-
+  @Autowired
+  private UserDomainService userDomainService;
+  private MessageSource messages;
   private final Logger apLog = Logger.getLogger(SystemPropertyConstants.APPLICATION_LOGGER);
 
   /**
    * 全てのユーザーを取得します。
    * 
-   * @return 全てのユーザー。
+   * @return 全てのユーザーのリスト。
    */
   public List<User> getUsers() {
-    apLog.info("全てのユーザーを取得します。");
+    apLog.info(messages.getMessage(MessageIdConstants.D_USER_GET_USERS, new Object[] {}, Locale.getDefault()));
     return userRepository.findAll();
   }
 
@@ -41,10 +48,10 @@ public class UserApplicationService {
    * @throws UserNotFoundException ユーザーが見つからない場合。
    */
   public User getUser(long id) throws UserNotFoundException {
-    apLog.info("ユーザーIDが、" + id + "のユーザーを取得します。");
+    apLog.info(messages.getMessage(MessageIdConstants.D_USER_GET_USER,
+        new Object[] { id }, Locale.getDefault()));
     User user = userRepository.findById(id);
     if (user == null) {
-      apLog.info("ユーザーIDが、" + id + "のユーザーが見つかりませんでした。");
       throw new UserNotFoundException(id);
     }
     return user;
@@ -55,10 +62,15 @@ public class UserApplicationService {
    * 
    * @param user 追加するユーザー。
    * @return 追加されたユーザー。
+   * @throws UserAlreadyExistException ユーザーが既に存在する場合。
    */
-  public User addUser(User user) {
-    final String userName = user.getUserName();
-    apLog.info(userName + "を追加します。");
+  public User addUser(User user) throws UserAlreadyExistException {
+    final long id = user.getId();
+    apLog.info(messages.getMessage(MessageIdConstants.D_USER_ADD_USER,
+        new Object[] { id }, Locale.getDefault()));
+    if (userDomainService.isExistUser(user)) {
+      throw new UserAlreadyExistException(id);
+    }
     User addedUser = userRepository.add(user);
     return addedUser;
   }
@@ -67,20 +79,31 @@ public class UserApplicationService {
    * ユーザーを更新します。
    * 
    * @param user 更新するユーザー。
+   * @throws UserNotFoundException ユーザーが見つからない場合。
    */
-  public void updateUser(User user) {
-    final String userName = user.getUserName();
-    apLog.info(userName + "を更新します。");
+  public void updateUser(User user) throws UserNotFoundException {
+    final long id = user.getId();
+    apLog.info(messages.getMessage(MessageIdConstants.D_USER_UPDATE_USER,
+        new Object[] { id }, Locale.getDefault()));
+    if (!userDomainService.isExistUser(user)) {
+      throw new UserNotFoundException(id);
+    }
     userRepository.update(user);
   }
 
   /**
    * ユーザーを削除します。
    * 
-   * @param id   ユーザーのID。
+   * @param id ユーザーのID。
+   * @throws UserNotFoundException ユーザーが見つからない場合。
    */
-  public void deleteUser(long id) {
-    apLog.info("ユーザーID: " + id + "を削除します。");
+  public void deleteUser(long id) throws UserNotFoundException {
+    apLog.info(messages.getMessage(MessageIdConstants.D_USER_DELETE_USER,
+        new Object[] { id }, Locale.getDefault()));
+    User user = userRepository.findById(id);
+    if (user == null) {
+      throw new UserNotFoundException(id);
+    }
     userRepository.delete(id);
   }
 }
